Fix Speech Commands test set url error
The code currently returns this error when attempting to load the test
set:
```
ValueError: unknown url type: 'speech_commands_test_set_v0.02'
```

This is indeed not a valid URL or a valid shorthand. Not even as far
back as torchaudio 0.7, so I am not sure if this ever worked.

The solution is to use the testing list from the main dataset.